Use post save locking to allow/disallow publishing
This addresses https://github.com/humanmade/altis-workflow/issues/71 however with the major caveat that it removes the step to prompt users when they are about to publish without completing all tasks when block on failing is set to false.

The checks are still visible right under the publish button and in some ways this looks better and will be more compatible with other plugins etc. Worth a discussion.