Move lock/unlock save button logic into icon component
The ChecklistPanel component is sometimes hidden in the
PrePublishSidebar, and when the save lock is only updated on rendering
of that component there are times when it doesn't unlock when it should.

This change fixes that issue by moving the lock/unlock logic into the
toolbar, which is always visible.